email verification route is protected
