--- backend/src/services/mail-verification/send.service.ts
@@ -1,10 +1,0 @@
-const mailSend = (email: string): string => {
-  // Mock token, it must be genereted here
-  const token = 'rOiBUBf2pPvOHTNBeTR8CkLGokwHADhv';
-  // TODO Please let me know what this function should do @RomanMahola @Rudych
-  // eslint-disable-next-line no-console
-  console.log(`Sending email to ${email}`);
-  return token;
-};
-
-export { mailSend };
