+++ frontend/src/components/mail-verification/verification-failed/styles.module.scss
@@ -10,6 +10,6 @@
 }
 
 .message {
-  margin-bottom: 30px;
+  margin-bottom: 28px;
   font-size: 16px;
 }
