+++ shared/src/validation-schemas/email/email-validation-schema.ts
@@ -1,0 +1,15 @@
+import * as Yup from 'yup';
+
+import { UserValidationMessage } from '../../common/enums/enums';
+
+const emailSchema = Yup.object().shape({
+  email: Yup.string()
+    .required(UserValidationMessage.EMAIL_REQUIRED)
+    .email(UserValidationMessage.INVALID_EMAIL)
+    .matches(/^[A-Za-z0-9.\-_]*@/, UserValidationMessage.INVALID_EMAIL)
+    .matches(/@[A-Za-z0-9.\-_]*$/, UserValidationMessage.INVALID_EMAIL)
+    .matches(/^[\S]{1,35}@/, UserValidationMessage.INVALID_EMAIL_LENGTH)
+    .matches(/@[\S]{3,35}$/, UserValidationMessage.INVALID_EMAIL_LENGTH),
+});
+
+export { emailSchema };
