+++ shared/src/validation-schemas/validation-schemas.ts
@@ -4,3 +4,4 @@ export {
   forgotPassword,
   resetPassword,
 } from './forgot-password/forgot-password.validation-schema.js';
+export { emailSchema } from './email/email-validation-schema';
